requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pillow
pymupdf
python-dotenv
openai
sqlalchemy
passlib
bcrypt
requests
pyhton-jose
jwt